requirements.txt
discord.py
python-dotenv
aiohttp
PyYAML
ujson
dotenv